requirements.txt
authlib ~= 1.0
django ~= 4.0
python-dotenv ~= 0.19
requests ~= 2.27
jsonfield ~= 3.1.0
gunicorn ~= 21.2.0
django-background-task ~= 0.1.8
django-background-tasks ~= 1.2.5
django-background-tasks-updated ~= 1.2.7
beautifulsoup4 ~= 4.12.2
reportlab ~= 4.0.7
sqlparse ~= 0.4.4
virtualenv ~= 20.24.5